+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/controller/BalanceController.java
@@ -3,9 +3,9 @@ package kr.hhplus.be.server.reservation.interfaces.web.controller;
 import kr.hhplus.be.server.reservation.application.balance.ChargeBalanceUseCase;
 import kr.hhplus.be.server.reservation.application.balance.GetBalanceUseCase;
 import kr.hhplus.be.server.reservation.domain.model.Balance;
-import kr.hhplus.be.server.reservation.interfaces.web.dto.request.BalanceChargeRequest;
-import kr.hhplus.be.server.reservation.interfaces.web.dto.response.BalanceChargeResponse;
-import kr.hhplus.be.server.reservation.interfaces.web.dto.response.WalletBalanceResponse;
+import kr.hhplus.be.server.reservation.interfaces.web.dto.request.balance.BalanceChargeRequest;
+import kr.hhplus.be.server.reservation.interfaces.web.dto.response.balance.BalanceChargeResponse;
+import kr.hhplus.be.server.reservation.interfaces.web.dto.response.balance.WalletBalanceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
