+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/controller/ReservationTokenController.java
@@ -1,11 +1,9 @@
 package kr.hhplus.be.server.reservation.interfaces.web.controller;
 
 import kr.hhplus.be.server.queue.dto.request.QueueTokenIssueRequest;
-import kr.hhplus.be.server.queue.dto.response.QueueStatusResponse;
-import kr.hhplus.be.server.queue.dto.response.QueueTokenIssueResponse;
 import kr.hhplus.be.server.reservation.application.reservationToken.CheckQueueStatusUseCase;
 import kr.hhplus.be.server.reservation.application.reservationToken.IssueReservationTokenUseCase;
-import kr.hhplus.be.server.reservation.interfaces.web.dto.response.ReservationTokenStatusResponse;
+import kr.hhplus.be.server.reservation.interfaces.web.dto.response.reservation.ReservationTokenStatusResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
